regenerate config.sub
This generates a fresh copy of config.sub.

The existing config.sub calls 64-bit ARM targets `aarch64`, which is
the official name for the target. However, Apple targets go by the
name `arm64-apple`. As a result, this platform was not recognized.

This gets iverilog building on Apple targets with 64-bit ARM processors.

Signed-off-by: Sean Cross <sean@xobs.io>